Captcha: move exceptions in sub namespace Exception
