+++ library/Zend/Captcha/Exception/NoFontProvidedException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Captcha;
+namespace Zend\Captcha\Exception;
+use Zend\Captcha\Exception;
 
 /**
  * Exception for Zend_Form component.
@@ -33,7 +34,7 @@ namespace Zend\Captcha;
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class ExtensionNotLoadedException
+class NoFontProvidedException
     extends \Exception
     implements Exception
 {
